Fix some bug and Integrate some code
